+++ tests/Models/ModelWithFlaggedQueries.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace BenSampo\Enum\Tests\Models;
+
+use BenSampo\Enum\Traits\FlaggedEnumQueries;
+use Illuminate\Database\Eloquent\Model;
+
+class ModelWithFlaggedQueries extends Model
+{
+    use FlaggedEnumQueries;
+
+    public $table = 'test_models';
+    
+    protected $guarded = [];
+}